test(scripts): extract indexnow ref parser and cover it
scripts/indexnow-changed-urls.mjs parsed its "category/slug" CLI ref args inline,
next to the fetch layer that expands each ref into hub URLs, so the ref parsing
was untestable without the network. Move the pure parser into
scripts/lib/indexnow-refs.mjs - joining the existing lib/indexnow* helpers the
script already imports - and import it back.

- scripts/lib/indexnow-refs.mjs: parseRefs.
- scripts/indexnow-changed-urls.mjs: import parseRefs; drop the local copy.
- tests: cover valid category/slug parsing with trimming, and the skips for
  blank refs, refs with no slash, empty category/slug, and nested-slug slugs,
  plus a mixed valid/invalid batch. Branch coverage 100% (9/9).

No behavior change; the emitted URL set is identical.
